% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bins_method.R
\name{bins_method}
\alias{bins_method}
\title{Calculate bin width based on specified method}
\usage{
bins_method(x, method = "sturges")
}
\arguments{
\item{x}{numeric vector of data points}

\item{method}{method for calculating bin width; options are "sturges", "scott", "fd", or a numeric value specifying the number of bins}
}
\value{
Numeric value representing the calculated bin width
}
\description{
Calculate the bin width for histogramming data based on different methods: Sturges, Scott, Freedman-Diaconis, or a user-defined number of bins.
}
\examples{

data <- rnorm(1000)
bins_method(data, method = "sturges")
bins_method(data, method = "scott")
bins_method(data, method = "fd")
bins_method(data, method = 30)  # User-defined number of bins

}
