% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcINO.R
\name{calcINO}
\alias{calcINO}
\title{Process raw eye-tracker data to determine presence of INO}
\usage{
calcINO(ETdata)
}
\arguments{
\item{ETdata}{Dataframe with raw eye-tracker pro-saccades data after processing according to DEMoNS protocol.}
}
\value{
Dataframe with additional columns indicating presence of INO based on AUC and PvAm metrics, as well as quality control metrics and VDI calculations.
}
\description{
Process raw eye-tracker data to determine presence of INO
}
