% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_scatterplots.R
\name{check_scatterplots}
\alias{check_scatterplots}
\title{check_scatterplots: Create scatterplots for outcome-predictor combinations with outliers labeled}
\usage{
check_scatterplots(data, outcomes, predictors, id)
}
\arguments{
\item{data}{dataframe containing the variables}

\item{outcomes}{vector of outcome variable names}

\item{predictors}{vector of predictor variable names}

\item{id}{variable used to identify each participant, e.g. pat_id}
}
\value{
A list of scatterplots for each outcome-predictor combination with outliers labeled.
}
\description{
check_scatterplots: Create scatterplots for outcome-predictor combinations with outliers labeled
}
\examples{

check_scatterplots(
data = MS_trial_data,
outcomes = c("EDSS", "SDMT"),
predictors = c("age", "time"),
id = "pat_id"
)

}
