% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cutoff_VDI}
\alias{cutoff_VDI}
\alias{cutoff_VDI_AUC}
\alias{cutoff_VDI_PvAm}
\title{VDI cutoff values for determining INO}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\source{
Cutoff values were established in Nij Bijvank et al. 2019 (https://doi.org/10.1212/WNL.0000000000007499).
}
\usage{
cutoff_VDI

cutoff_VDI_AUC

cutoff_VDI_PvAm
}
\description{
Dataframe with cutoff values for versional dysconjugacy indices (VDI) for area under the curve (VDI-AUC) and peak velocity corrected for amplitude (VDI-pV/Am) to determine presence of internuclear ophthalmoplegia (INO).
VDI data should be obtained using prosaccades task from the DEMoNS protocol (Nij Bijvank et al. 2018, https://doi.org/10.1371/journal.pone.0200695).
Cutoff values were established in Nij Bijvank et al. 2019 (https://doi.org/10.1212/WNL.0000000000007499).
}
\keyword{datasets}
