% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEMoNS_data}
\alias{DEMoNS_data}
\alias{DEMoNS_data_fix}
\alias{DEMoNS_data_prosac}
\alias{DEMoNS_data_antisac}
\alias{DEMoNS_data_expsac}
\alias{DEMoNS_data_dblstep}
\alias{DEMoNS_data_repsac}
\title{Example measurement DEMoNS eyetracking protocol}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 918 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 105 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 219 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 96 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 99 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 310 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 99 columns.
}
\source{
Generated using tasks and scripts available at \href{https://www.protocols.io/view/demons-protocol-for-measurement-and-analysis-of-ey-x54v98eyml3e/v3}{protocols.io}.
}
\usage{
DEMoNS_data

DEMoNS_data_fix

DEMoNS_data_prosac

DEMoNS_data_antisac

DEMoNS_data_expsac

DEMoNS_data_dblstep

DEMoNS_data_repsac
}
\description{
Example data obtained from two measurements using the DEMoNS eyetracking protocol \href{https://doi.org/10.1371/journal.pone.0200695}{(Nij Bijvank et al. 2018)}. Task and analysis MATLAB scripts are available at \href{https://www.protocols.io/view/demons-protocol-for-measurement-and-analysis-of-ey-x54v98eyml3e/v3}{protocols.io}.
Includes data for fixation, pro-saccades, anti-saccades, express saccades, double-step saccades and repeated saccades task.
For data specific to single tasks, see \link{DEMoNS_data_fix}, \link{DEMoNS_data_prosac}, \link{DEMoNS_data_antisac}, \link{DEMoNS_data_expsac}, \link{DEMoNS_data_dblstep} and \link{DEMoNS_data_repsac}.
}
\keyword{datasets}
