% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demons_process_dblstep.R
\name{demons_process_dblstep}
\alias{demons_process_dblstep}
\title{Process double-step saccades task data from the DEMoNS eye-tracking protocol}
\usage{
demons_process_dblstep(
  demons_data_dblstep,
  keep_essential = FALSE,
  add_prefix = FALSE
)
}
\arguments{
\item{demons_data_dblstep}{Dataframe with raw eye-tracker double-step saccades data after processing according to DEMoNS protocol.}

\item{keep_essential}{Logical indicating whether to only keep essential identifier and outcome columns (default = FALSE).}

\item{add_prefix}{Logical indicating whether to add "fix_" prefix to all columns except identifier columns (default = FALSE).}
}
\value{
Dataframe with additional columns indicating quality control metrics and calculated double-step saccade parameters.
}
\description{
Calculates quality control metrics and additional parameters for double-step saccades task data obtained using the DEMoNS eye-tracking protocol.
Allows selection of essential variables and addition of a prefix to variable names.
See \href{https://doi.org/10.1093/brain/awac474}{Nij Bijvank et al. 2023} for details on double-step saccade parameters.
See \href{https://doi.org/10.1371/journal.pone.0200695}{Nij Bijvank et al. 2018} for protocol details.
DEMoNS protocol task and MATLAB analysis script available at \href{https://www.protocols.io/view/demons-protocol-for-measurement-and-analysis-of-ey-x54v98eyml3e/v3}{protocols.io}.
}
\examples{
demons_process_dblstep(DEMoNS_data_dblstep, keep_essential = TRUE, add_prefix = TRUE)

}
