% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demons_process_expsac.R
\name{demons_process_expsac}
\alias{demons_process_expsac}
\title{Process Express saccades task data from the DEMoNS eye-tracking protocol}
\usage{
demons_process_expsac(
  demons_data_expsac,
  keep_essential = FALSE,
  add_prefix = FALSE
)
}
\arguments{
\item{demons_data_expsac}{Dataframe with raw eye-tracker express saccades data after processing according to DEMoNS protocol.}

\item{keep_essential}{Logical indicating whether to only keep essential identifier and outcome columns (default = FALSE).}

\item{add_prefix}{Logical indicating whether to add "fix_" prefix to all columns except identifier columns (default = FALSE).}
}
\value{
Dataframe with additional columns indicating quality control metrics.
}
\description{
Calculates quality control metrics for express saccades task data obtained using the DEMoNS eye-tracking protocol.
Allows selection of essential variables and addition of a prefix to variable names.
See \href{https://doi.org/10.1371/journal.pone.0200695}{Nij Bijvank et al. 2018} for protocol details.
DEMoNS protocol task and MATLAB analysis script available at \href{https://www.protocols.io/view/demons-protocol-for-measurement-and-analysis-of-ey-x54v98eyml3e/v3}{protocols.io}.
}
\examples{
demons_process_expsac(DEMoNS_data_expsac, keep_essential = TRUE, add_prefix = TRUE)

}
