% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demons_process_fix.R
\name{demons_process_fix}
\alias{demons_process_fix}
\title{Process fixation task data from the DEMoNS eye-tracking protocol}
\usage{
demons_process_fix(demons_data_fix, keep_essential = FALSE, add_prefix = FALSE)
}
\arguments{
\item{demons_data_fix}{Dataframe with fixation task data obtained using the DEMoNS eye-tracking protocol.}

\item{keep_essential}{Logical indicating whether to only keep essential identifier and outcome columns (default = FALSE).}

\item{add_prefix}{Logical indicating whether to add "fix_" prefix to all columns except identifier columns (default = FALSE).}
}
\value{
Dataframe with additional columns for fixation task parameters and dichotomous variables based on established cutoffs.
}
\description{
Calculates additional parameters and dichotomous variables based on established cutoffs for fixation task data obtained using the DEMoNS eye-tracking protocol.
Can additionally only retain essential columns and/or add a prefix to all columns except identifier columns.
See \href{https://doi.org/10.1167/iovs.18-26096}{Nij Bijvank et al. 2019} for cutoff references.
See \href{https://doi.org/10.1371/journal.pone.0200695}{Nij Bijvank et al. 2018} for protocol details.
DEMoNS protocol task and MATLAB analysis script available at \href{https://www.protocols.io/view/demons-protocol-for-measurement-and-analysis-of-ey-x54v98eyml3e/v3}{protocols.io}.
}
\examples{
demons_process_fix(DEMoNS_data_fix, keep_essential = TRUE, add_prefix = TRUE)

}
