% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demons_process_INO.R
\name{demons_process_INO}
\alias{demons_process_INO}
\title{Process DEMoNS protocol eye-tracker data to calculate INO parameters}
\usage{
demons_process_INO(demons_data_prosac)
}
\arguments{
\item{demons_data_prosac}{Dataframe with raw eye-tracker pro-saccades data after processing according to DEMoNS protocol.}
}
\value{
Dataframe with additional columns indicating presence of INO based on AUC and PvAm metrics, as well as quality control metrics and VDI calculations.
}
\description{
Process DEMoNS protocol eye-tracker data to calculate INO parameters
}
