% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demons_process_prosac.R
\name{demons_process_prosac}
\alias{demons_process_prosac}
\title{Process pro-saccades task data from the DEMoNS eye-tracking protocol}
\usage{
demons_process_prosac(
  demons_data_prosac,
  keep_essential = FALSE,
  add_prefix = FALSE
)
}
\arguments{
\item{demons_data_prosac}{Dataframe with raw eye-tracker pro-saccades data after processing according to DEMoNS protocol.}

\item{keep_essential}{Logical indicating whether to only keep essential identifier and outcome columns (default = FALSE).}

\item{add_prefix}{Logical indicating whether to add "fix_" prefix to all columns except identifier columns (default = FALSE).}
}
\value{
Dataframe with additional columns indicating presence of INO based on AUC and PvAm metrics, as well as quality control metrics and VDI calculations.
}
\description{
Calculates INO parameters and quality control metrics for pro-saccades task data obtained using the DEMoNS eye-tracking protocol.
Allows selection of essential variables and addition of a prefix to variable names.
See also \code{\link[=demons_process_INO]{demons_process_INO()}} for INO calculation details.
See \href{https://doi.org/10.1371/journal.pone.0200695}{Nij Bijvank et al. 2018} for protocol details.
DEMoNS protocol task and MATLAB analysis script available at \href{https://www.protocols.io/view/demons-protocol-for-measurement-and-analysis-of-ey-x54v98eyml3e/v3}{protocols.io}.
}
\examples{
demons_process_prosac(DEMoNS_data_prosac, keep_essential = TRUE, add_prefix = TRUE)
}
