% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeglm_log_mult.R
\name{geeglm_log_mult}
\alias{geeglm_log_mult}
\title{geeglm_log_mult: run multiple logistic GEE regressions and output in table}
\usage{
geeglm_log_mult(
  data,
  outcomes,
  predictors,
  covariates = "",
  id,
  corstr = "exchangeable",
  exponentiate = TRUE
)
}
\arguments{
\item{data}{data frame in long format with a row for every observation of the outcome variables. See \link{MS_trial_data} for an example dataset.}

\item{outcomes}{a vector with your outcome variables (dependent variables), where the names of your variables are in quotes. For example: c(‘outcome1’, ‘outcome2’, ‘outcome3’).}

\item{predictors}{A vector containing your predictors (independent variables), where the names of your variables are in quotes. For example: c(‘predictor1’, ‘predictor2’, ‘predictor3’). It is possible to directly recode binary or categorical variables and get an output for both encodings, e.g. c(‘predictor1’, ‘fct_rev(predictor1)’). It is also possible to specify interaction terms, e.g. c(‘predictor1 * predictor2’, ‘predictor2 * predictor 3’).}

\item{covariates}{A vector containing your covariates (variables you want to correct for), where the entire enumeration of covariates per model are in quotes and different covariates in the same model are separated by a plus. Also, covariates should be preceded by a plus. For example: c(‘+ covariate 1 + covariate 2’, ‘+ covariate2 + covariate3’). If you want to run a model with and without covariates then specify it like this: c(‘’, ‘+ covariate 1 + covariate 2’)}

\item{id}{Variable used to identify each participant, e.g. pat_id.}

\item{corstr}{The correlation structure for your GEE analysis. Default is "echangeable". See ?geeglm \code{\link[geepack:geeglm]{geepack::geeglm()}} for more information.}

\item{exponentiate}{Do you want to exponentiate the output (e^x) so that you get the odds ratio? Default is TRUE.}
}
\value{
Dataframe with results from multiple logistic GEE regression analyses.
}
\description{
geeglm_log_mult: run multiple logistic GEE regressions and output in table
}
\examples{

geeglm_log_mult(
data = MS_trial_data,
outcomes = "INO",
predictors = c("intervention", "intervention * time"),
covariates = c("", "+ gender + age"),
id = "pat_id"
)

}
