% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeglm_log_mult_f2m.R
\name{geeglm_log_mult_f2m}
\alias{geeglm_log_mult_f2m}
\title{geeglm_log_mult_f2m: Formula dataframe to dataframe with logistic GEE models}
\usage{
geeglm_log_mult_f2m(
  df_formulas,
  data,
  id,
  corstr = "exchangeable",
  progress = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{df_formulas}{Dataframe created with \link[=construct_formulas]{construct_formulas()} containing formulas for regression analyses.}

\item{data}{data frame in long format with a row for every observation of the outcome variables. See \link{MS_trial_data} for an example dataset.}

\item{id}{Variable used to identify each participant, e.g. pat_id. Internally coerced to a factor for model fitting.}

\item{corstr}{The correlation structure for your GEE analysis. Default is "echangeable". See ?geeglm \code{\link[geepack:geeglm]{geepack::geeglm()}} for more information.}

\item{progress}{Whether to show a progress bar. Use TRUE to turn on a progress bar, default is FALSE.}

\item{quiet}{Whether to suppress error messages from model fitting. Default is FALSE. Error messages are always captured and stored in the output dataframe.}
}
\value{
Dataframe provided as "df_formulas" with appended columns containing logistic GEE models and error messages.
}
\description{
Run and append logistic GEE models based on a formula dataframe created with \link[=construct_formulas]{construct_formulas()}.
The output can be used as input to \link[=geeglm_log_mult_m2p]{geeglm_log_mult_m2p()} to extract parameters in a tidy format.
}
\examples{
df_formulas <- construct_formulas(
outcomes = "INO",
predictors = c("intervention", "intervention * time"),
covariates = c("", "+ gender + age")
)

geeglm_log_mult_f2m(df_formulas, data = MS_trial_data, id = "pat_id")

}
