% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeglm_log_mult_m2p.R
\name{geeglm_log_mult_m2p}
\alias{geeglm_log_mult_m2p}
\title{geeglm_log_mult_m2p: extract parameters from multiple logistic GEE models to tidy table}
\usage{
geeglm_log_mult_m2p(df_reg, exponentiate = TRUE, progress = FALSE)
}
\arguments{
\item{df_reg}{Dataframe output from \link[=geeglm_log_mult_f2m]{geeglm_log_mult_f2m()}.}

\item{exponentiate}{Do you want to exponentiate the output (e^x) so that you get the odds ratio? Default is TRUE.}

\item{progress}{Whether to show a progress bar. Use TRUE to turn on a progress bar, default is FALSE.}
}
\value{
Dataframe with results from multiple logistic GEE regression analyses.
}
\description{
Extract parameters from multiple logistic GEE models in a dataframe into a tidy table.
}
\examples{
df_formulas <- construct_formulas(
outcomes = "INO",
predictors = c("intervention", "intervention * time"),
covariates = c("", "+ gender + age")
)

df_reg <- geeglm_log_mult_f2m(df_formulas, data = MS_trial_data, id = "pat_id")

geeglm_log_mult_m2p(df_reg)

}
