% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_extensions.R
\name{ggsavemagnify}
\alias{ggsavemagnify}
\title{ggsave with magnification option}
\usage{
ggsavemagnify(filename, width, height, units = "cm", magnify, dpi)
}
\arguments{
\item{filename}{File path and name to save the plot.}

\item{width}{Width of plot}

\item{height}{Height of plot}

\item{units}{Units of width and height. Default is "cm".}

\item{magnify}{Magnification factor to scale down the width and height for saving.}

\item{dpi}{Dots per inch for the saved plot.}
}
\value{
Saves the current ggplot to a file with specified dimensions adjusted by the magnification factor.
}
\description{
ggsave with magnification option
}
