% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_extensions.R
\name{ggsavepp}
\alias{ggsavepp}
\alias{ggsavepphalf}
\alias{ggsaveppfull}
\alias{ggsaveppquart}
\title{ggsave with defaults for exporting to powerpoint}
\usage{
ggsavepp(filename)

ggsavepphalf(filename)

ggsaveppfull(filename)

ggsaveppquart(filename)
}
\arguments{
\item{filename}{File path and name to save the plot.}
}
\value{
Saves the current ggplot to a file with specified dimensions suitable for PowerPoint presentations.
}
\description{
ggsave with defaults for exporting to powerpoint
}
