% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_log_mult_f2m.R
\name{glm_log_mult_f2m}
\alias{glm_log_mult_f2m}
\title{glm_log_mult_f2m: Formula dataframe to dataframe with logistic regression models}
\usage{
glm_log_mult_f2m(
  df_formulas,
  data,
  std_models = FALSE,
  progress = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{df_formulas}{Dataframe containing regression formulas in a column named 'formula', created with \code{\link[=construct_formulas]{construct_formulas()}}.}

\item{data}{dataframe where each row represents an observation of the outcome variable.}

\item{std_models}{Do you want to run standardized versions of the models as well? This allows for standardized log(odds). Default is FALSE. For logistic regression only predictors and covariates are standardized as outcome is already on the same scale (OR). Be careful with interpretation.}

\item{progress}{Whether to show a progress bar. Use TRUE to turn on a progress bar, default is FALSE.}

\item{quiet}{Whether to suppress error messages from model fitting. Default is FALSE. Error messages are always captured and stored in the output dataframe.}
}
\value{
Dataframe provided as "df_formulas" with appended columns containing logistic regression models and error messages.
}
\description{
Run and append logistic regression models based on a formula dataframe created with \code{\link[=construct_formulas]{construct_formulas()}}.
The output can be used as input to \code{\link[=glm_log_mult_m2p]{glm_log_mult_m2p()}} to extract parameters in a tidy format.
}
\examples{

df_formulas <- construct_formulas(
outcomes = "INO",
predictors = c("intervention", "age"),
covariates = c("", "+ gender")
)

glm_log_mult_f2m(df_formulas, data = MS_trial_data)

}
