% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_log_mult_m2p.R
\name{glm_log_mult_m2p}
\alias{glm_log_mult_m2p}
\title{glm_log_mult_m2p: extract parameters from multiple logistic regression models to tidy table}
\usage{
glm_log_mult_m2p(df_reg, exponentiate = TRUE, progress = FALSE)
}
\arguments{
\item{df_reg}{Dataframe output from \code{\link[=glm_log_mult_f2m]{glm_log_mult_f2m()}}.}

\item{exponentiate}{Do you want to exponentiate the output (e^x) so that you get the odds ratio? Default is TRUE.}

\item{progress}{Whether to show a progress bar. Use TRUE to turn on a progress bar, default is FALSE.}
}
\value{
Dataframe with results from multiple logistic regression analyses.
}
\description{
glm_log_mult_m2p: extract parameters from multiple logistic regression models to tidy table
}
\examples{
df_formulas <- construct_formulas(
outcomes = "INO",
predictors = c("intervention", "age"),
covariates = c("", "+ gender")
)

df_reg <- glm_log_mult_f2m(df_formulas, data = MS_trial_data)

glm_log_mult_m2p(df_reg)

}
