% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_sp.R
\name{histogram_sp}
\alias{histogram_sp}
\title{Histogram with mean, median, SD, boxplot and outliers}
\usage{
histogram_sp(
  data,
  var,
  id = dplyr::row_number(),
  bins_method = "sturges",
  facet_cols = NULL,
  facet_rows = NULL,
  max.overlaps = 20
)
}
\arguments{
\item{data}{Dataframe containing numeric variables to check for normality.}

\item{var}{Numeric variable to create histogram for.}

\item{id}{Identifier for labeling outliers in plots, default is row number.}

\item{bins_method}{Method for determining histogram bin width. Choices are "sturges", "scott", "fd" (Freedman-Diaconis) or a numeric value specifying the number of bins. Default is "sturges".}

\item{facet_cols}{Variable to split data by in rows for faceted plots.}

\item{facet_rows}{Variable to split data by in columns for faceted plots.}

\item{max.overlaps}{Maximum number of overlaps allowed for outlier labels before suppression. Default is 20. Use "Inf" to disable suppression entirely.}
}
\value{
ggplot2 histogram object with outliers labeled.
}
\description{
Histogram with mean, median, SD, boxplot and outliers
}
\examples{

histogram_sp(
data = MS_trial_data,
var = SDMT,
id = pat_id,
facet_cols = intervention,
facet_rows = INO
)

}
