% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{is_outlier: Determine outliers according to Tukey method}
\usage{
is_outlier(x)
}
\arguments{
\item{x}{Numeric vector to check for outliers.}
}
\value{
Logical vector indicating whether each element in \code{x} is an outlier (TRUE) or not (FALSE).
}
\description{
This function identifies whether a value is an outlier in a numeric vector based on the Tukey method, which defines outliers as values that fall below Q1 - 1.5\emph{IQR or above Q3 + 1.5}IQR.
}
\examples{
data_vector <- c(10, 12, 14, 15, 18, 20, 22, 100)
is_outlier(data_vector)
}
