% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_assump_tests.R
\name{lm_assump_tests}
\alias{lm_assump_tests}
\title{lm_assump_tests: Generate plots for checking linear regression assumptions}
\usage{
lm_assump_tests(model)
}
\arguments{
\item{model}{A linear model object created with \code{lm()}.}
}
\value{
Plots for checking linear regression assumptions: a histogram of residuals for normality, a residuals vs fitted plot for linearity, and a scale-location plot for homoscedasticity.
}
\description{
Generate plots for checking linear regression assumptions: normality of residuals, linearity, and homoscedasticity.
}
\examples{

lm(SDMT ~ age, data = MS_trial_data) -> model
lm_assump_tests(model)

}
