% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mult.R
\name{lm_mult}
\alias{lm_mult}
\title{lm_mult: Run multiple linear regression models and output table}
\usage{
lm_mult(
  data,
  outcomes,
  predictors,
  covariates = "",
  formulas = NULL,
  std_beta = FALSE,
  exp_log = FALSE,
  progress = FALSE
)
}
\arguments{
\item{data}{data frame where each row represents an observation of the outcome variable.}

\item{outcomes}{a vector with outcome variables (dependent variables), where the names of the outcome variables are in quotes, e.g. \code{c("outcome1", "outcome2", "outcome3")}. It is possible to specify log transformation of the outcome variables, e.g. \code{c("log(outcome1)", "log(outcome2)")}.}

\item{predictors}{a vector with predictors (independent variables), where the names of the variables are in quotes, e.g. \code{c("predictor1", "predictor2", "predictor3")}. It is possible to directly recode binary or categorical variables and get an output for both encodings, e.g. \code{c("predictor1", "fct_rev(predictor1)")}. It is also possible to specify interaction terms, e.g. \code{c("predictor1 * predictor2", "predictor2 * predictor 3")}.}

\item{covariates}{a vector containing covariates (variables you want to correct for), where the entire enumeration of covariates are in quotes, different covariates in the same model are separated and preceded by a plus, e.g. \code{c("+ covariate 1 + covariate 2", "+ covariate2 + covariate3")}. If you want to run a model with and without covariates then specify it like this: \code{c("", "+ covariate 1 + covariate 2")}.}

\item{formulas}{optional vector of formulas specified manually, e.g. \code{c("outcome1 ~ predictor1 + covariate1 + (1|var1)", "outcome2 ~ predictor2 + covariate2 + (1|var2)")}. You can combine this with using the other inputs for single variables, all models are then run.}

\item{std_beta}{Do you also want standardized betas in the output, set this to TRUE.}

\item{exp_log}{Do you want to exponentiate the results for log transformed outcomes, set this to TRUE.}

\item{progress}{Whether to show a progress bar. Use TRUE to turn on a progress bar, default is FALSE.}
}
\value{
Large table with results from all regression formulas
}
\description{
These functions allow you to specify multiple outcome measures, predictors and covariates.
The function then creates regression formulas from all combinations of these variables.
The output is a large table with results from all regression formulas.
}
\examples{
lm_mult(
data = MS_trial_data,
outcomes = c("EDSS", "SDMT"),
predictors = "age",
covariates = c("", " + gender")
)

}
