% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mult_assump_tests.R
\name{lm_mult_assump_tests}
\alias{lm_mult_assump_tests}
\title{lm_mult_assump_tests: Generate plots for checking linear regression assumptions for multiple models}
\usage{
lm_mult_assump_tests(data, outcomes, predictors, covariates = "")
}
\arguments{
\item{data}{data frame with all your data.}

\item{outcomes}{a vector with your outcome variables (dependent variables), where the names of your variables are in quotes. For example: c(‘outcome1’, ‘outcome2’, ‘outcome3’).}

\item{predictors}{A vector containing your predictors (independent variables), where the names of your variables are in quotes. For example: c(‘predictor1’, ‘predictor2’, ‘predictor3’). It is possible to directly recode binary or categorical variables and get an output for both encodings, e.g. c(‘predictor1’, ‘fct_rev(predictor1)’). It is also possible to specify interaction terms, e.g. c(‘predictor1 * predictor2’, ‘predictor2 * predictor 3’).}

\item{covariates}{a vector containing your covariates (variables you want to correct for), where the entire enumeration of covariates per model are in quotes and different covariates in the same model are separated by a plus. Also, covariates should be preceded by a plus. For example: c(‘+ covariate 1 + covariate 2’, ‘+ covariate2 + covariate3’). If you want to run a model with and without covariates then specify it like this: c(‘’, ‘+ covariate 1 + covariate 2’)}
}
\value{
A list of plots for checking linear regression assumptions for each model.
}
\description{
Runs multiple linear regression models, similar to \code{\link[=lm_mult]{lm_mult()}}, and passes each model to \code{\link[=lm_assump_tests]{lm_assump_tests()}} to generate plots for checking normality of residuals, linearity, and homoscedasticity.
}
\examples{

lm_mult_assump_tests(
data = MS_trial_data,
outcomes = c("EDSS", "SDMT"),
predictors = "age",
covariates = c("", " + gender")
)

}
