% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mult_f2m.R
\name{lm_mult_f2m}
\alias{lm_mult_f2m}
\title{lm_mult_f2m: Formula dataframe to dataframe with linear regression models}
\usage{
lm_mult_f2m(df_formulas, data, progress = FALSE)
}
\arguments{
\item{df_formulas}{Dataframe containing regression formulas in a column named 'formula', created with \code{\link[=construct_formulas]{construct_formulas()}}.}

\item{data}{dataframe where each row represents an observation of the outcome variable.}

\item{progress}{Whether to show a progress bar. Use TRUE to turn on a progress bar, default is FALSE.}
}
\value{
Datafram provided as "df_formulas" with appended columns containing linear regression models.
}
\description{
Run and append linear regression models based on a formula dataframe created with \code{\link[=construct_formulas]{construct_formulas()}}.
}
\examples{
df_formulas <- construct_formulas(
outcomes = c("EDSS", "SDMT"),
predictors = "age",
covariates = c("", "+ gender")
)

lm_mult_f2m(
df_formulas = df_formulas,
data = MS_trial_data
)

}
