% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mult_m2p.R
\name{lm_mult_m2p}
\alias{lm_mult_m2p}
\title{lm_mult_m2p: Extract parameters from multiple linear models in a dataframe into a tidy table}
\usage{
lm_mult_m2p(df_reg, exp_log = FALSE, progress = FALSE)
}
\arguments{
\item{df_reg}{Dataframe containing multiple linear models in a column named 'model' and, optionally standardized models in a column named "std_model". This dataframe can be created with \code{\link[=lm_mult_f2m]{lm_mult_f2m()}}.}

\item{exp_log}{Do you want to exponentiate the results for log transformed outcomes, set this to TRUE. Default is FALSE.}

\item{progress}{Whether to show a progress bar. Use TRUE to turn on a progress bar, default is FALSE.}
}
\value{
Dataframe with results from multiple linear regression analyses.
}
\description{
lm_mult_m2p: Extract parameters from multiple linear models in a dataframe into a tidy table
}
\examples{
df_formulas <- construct_formulas(
outcomes = c("EDSS", "SDMT"),
predictors = "age",
covariates = c("", "+ gender")
)

df_reg <- lm_mult_f2m(df_formulas, data = MS_trial_data)

lm_mult_m2p(df_reg)


}
