% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_mult_f2m.R
\name{lmer_mult_f2m}
\alias{lmer_mult_f2m}
\title{lmer_mult: Run multiple linear mixed effects models and output table}
\usage{
lmer_mult_f2m(
  df_formulas,
  data,
  REML = FALSE,
  control = lme4::lmerControl(),
  progress = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{data}{data frame with outcomes, predictors, covariates and variables for random effects in long format with a row for every observation of the outcome variables. See \link{MS_trial_data} for an example dataset.}

\item{REML}{logical scalar - Should the estimates be chosen to optimize the REML criterion (as opposed to the log-likelihood)?}

\item{control}{This gives you access to the ‘control’ input of the \code{\link[lme4:lmer]{lme4::lmer()}} function and \code{\link[lme4:lmerControl]{lme4::lmerControl()}}. This allows you to bypass certain errors. For example, in case of error about number of observations, you can try control = lmerControl(check.nobs.vs.nRE = ‘ignore’).}

\item{progress}{Whether to show a progress bar. Use TRUE to turn on a progress bar, default is FALSE.}

\item{quiet}{Whether to suppress error messages from model fitting. Default is FALSE. Error messages are always captured and stored in the output dataframe.}
}
\value{
Dataframe with linear mixed effects models appended as a list-column in the "model" column and error messages in the "model_error" column.
}
\description{
Run multiple linear mixed effect models with the lme4 and lmerTest packages.
}
\details{
"std_models" to obtain standardized regression coefficients may be implemented in the future, but is currently not available.
}
\examples{
df_formulas <- construct_formulas(
 outcomes = c("EDSS", "SDMT"),
 predictors = c("intervention", "intervention * time"),
 covariates = c("", "+ age + gender"),
 randoms = c("+ (1|pat_id)")
 )

 lmer_mult_f2m(df_formulas, data = MS_trial_data)

}
