% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_mult_m2p.R
\name{lmer_mult_m2p}
\alias{lmer_mult_m2p}
\title{lmer_mult_m2p: Extract parameters from multiple linear mixed effect models and output table}
\usage{
lmer_mult_m2p(df_reg, exp_log = FALSE, progress = FALSE)
}
\arguments{
\item{df_reg}{Dataframe created with \code{\link[=lmer_mult_f2m]{lmer_mult_f2m()}} containing the lmer models to be processed in the column "model"}

\item{exp_log}{Do you want to exponentiate the results for log transformed outcomes, set this to TRUE. Default is FALSE.}

\item{progress}{Whether to show a progress bar. Use TRUE to turn on a progress bar, default is FALSE.}
}
\value{
Large table with results from all regression formulas
}
\description{
Extract parameters from dataframe with multiple linear mixed models (i.e. \code{\link[lmerTest:lmer]{lmerTest::lmer()}}) generated with \code{\link[=lmer_mult_f2m]{lmer_mult_f2m()}}.
}
\details{
The output contains some additional variables that are not in the standard mixed model output namely:

isSingular = is TRUE if the model is ‘singular’. There is debate whether you can then use the model or not. See ?isSingular

Loglik = the log likelihood of the model with which to compare models

Deviance = -2 * loglik

term_ranova = output of the ranova() function that performs the likelihood ratio test(LRT). This variable indicates for which term in the regression formula the LRT was performed.

npar_ranova = output of the ranova() function performing the likelihood ratio test(LRT). This variable indicates the number of parameters in the model.

p_ranova = output of the ranova() function performing the likelihood ratio test(LRT). This variable indicates the p value of the LRT.

"std_beta" to obtain standardized regression coefficients may be implemented in the future, but is currently not available.
}
\examples{

df_formulas <- construct_formulas(
  outcomes = c("EDSS", "SDMT"),
  predictors = c("intervention", "intervention * time"),
  covariates = c("", "+ age + gender"),
  randoms = c("+ (1|pat_id)")
)

df_reg <- lmer_mult_f2m(df_formulas, data = MS_trial_data)

lmer_mult_m2p(df_reg)
}
