% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MS_trial_data}
\alias{MS_trial_data}
\title{Randomly generated MS trial data}
\format{
\subsection{'MS_trial_data'}{

A data frame with 500 rows and 8 columns:
\describe{
\item{pat_id}{Patient ID ranging from pat001 to pat100}
\item{age}{Baseline age ranging from 18 to 70 years with a mean of 40 years}
\item{intervention}{Patient in intervention or control group}
\item{gender}{Male or female}
\item{time}{Timepoint ranging from 1 to 5}
\item{INO}{Does patient have INO? INO is unlikely to resolve once patients have it}
\item{SDMT}{Score on the Symbol Digits Modalities test. Ranging from 0 to 60. Score is more likely to decline over time.}
\item{EDSS}{Expanded Disability Status Scale score. Ranging from 0 to 10 in increments of 0.5. Follows a bimodal distribution with peaks around 3 and 6. EDSS is more likely to increase over time.}
}
}
}
\source{
Generated using internal script.
}
\usage{
MS_trial_data
}
\description{
Randomly generated data from a fictitious MS trial for testing purposes. Includes 100 patients with 5 timepoints each.
}
\keyword{datasets}
