% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samspack-package.R
\docType{package}
\name{samspack-package}
\alias{samspack}
\alias{samspack-package}
\title{samspack: Sam's personal collection of functions}
\description{
Sam's personal collection of functions for data analysis of eye-tracking data and multiple sclerosis.
}
\author{
\strong{Maintainer}: Sam Hof \email{s.n.hof@amsterdamumc.nl} (\href{https://orcid.org/0000-0002-0453-6449}{ORCID})

}
\keyword{internal}
