% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFQ_scoring.R
\name{VFQ_scoring}
\alias{VFQ_scoring}
\title{VFQ-25 questionnaire data naming and scoring}
\usage{
VFQ_scoring(df)
}
\arguments{
\item{df}{dataframe with raw VFQ-25 data}
}
\value{
dataframe with added columns
}
\description{
VFQ-25 questionnaire data naming and scoring
}
