% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_histograms.R
\name{check_histograms}
\alias{check_histograms}
\title{check_histograms: Check normality for multiple variables in dataframe}
\usage{
check_histograms(
  data,
  facets_rows = NULL,
  facets_cols = NULL,
  id = dplyr::row_number(),
  bins_method = "sturges"
)
}
\arguments{
\item{data}{Dataframe containing numeric variables to check for normality.}

\item{facets_rows}{Variable to split data by in rows for faceted plots.}

\item{facets_cols}{Variable to split data by in columns for faceted plots.}

\item{id}{Identifier for labeling outliers in plots, default is row number.}

\item{bins_method}{Method for determining histogram bin width. Choices are "sturges", "scott", "fd" (Freedman-Diaconis) or a numeric value specifying the number of bins. Default is "sturges".}
}
\value{
List of ggplot2 histogram objects for each numeric variable, with outliers labeled.
}
\description{
check_histograms: Check normality for multiple variables in dataframe
}
\examples{
check_histograms(data = MS_trial_data, id = pat_id, facets_cols = gender)

}
