% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geeglm_log_mult.R
\name{geeglm_log_mult}
\alias{geeglm_log_mult}
\title{geeglm_log_mult: run multiple logistic GEE regressions and output in table}
\usage{
geeglm_log_mult(
  data,
  outcomes,
  predictors,
  covariates = "",
  formulas = NULL,
  id,
  corstr = "exchangeable",
  exponentiate = TRUE,
  progress = FALSE
)
}
\arguments{
\item{data}{data frame in long format with a row for every observation of the outcome variables. See \link{MS_trial_data} for an example dataset.}

\item{outcomes}{a vector with outcome variables (dependent variables), where the names of the outcome variables are in quotes, e.g. \code{c("outcome1", "outcome2", "outcome3")}. It is possible to specify log transformation of the outcome variables, e.g. \code{c("log(outcome1)", "log(outcome2)")}.}

\item{predictors}{a vector with predictors (independent variables), where the names of the variables are in quotes, e.g. \code{c("predictor1", "predictor2", "predictor3")}. It is possible to directly recode binary or categorical variables and get an output for both encodings, e.g. \code{c("predictor1", "fct_rev(predictor1)")}. It is also possible to specify interaction terms, e.g. \code{c("predictor1 * predictor2", "predictor2 * predictor 3")}.}

\item{covariates}{a vector containing covariates (variables you want to correct for), where the entire enumeration of covariates are in quotes, different covariates in the same model are separated and preceded by a plus, e.g. \code{c("+ covariate 1 + covariate 2", "+ covariate2 + covariate3")}. If you want to run a model with and without covariates then specify it like this: \code{c("", "+ covariate 1 + covariate 2")}.}

\item{formulas}{optional vector of formulas specified manually, e.g. \code{c("outcome1 ~ predictor1 + covariate1 + (1|var1)", "outcome2 ~ predictor2 + covariate2 + (1|var2)")}. You can combine this with using the other inputs for single variables, all models are then run.}

\item{id}{Variable used to identify each participant, e.g. pat_id.}

\item{corstr}{The correlation structure for your GEE analysis. Default is "echangeable". See ?geeglm \code{\link[geepack:geeglm]{geepack::geeglm()}} for more information.}

\item{exponentiate}{Do you want to exponentiate the output (e^x) so that you get the odds ratio? Default is TRUE.}

\item{progress}{Whether to show a progress bar. Use TRUE to turn on a progress bar, default is FALSE.}
}
\value{
Dataframe with results from multiple logistic GEE regression analyses.
}
\description{
geeglm_log_mult: run multiple logistic GEE regressions and output in table
}
\examples{

geeglm_log_mult(
data = MS_trial_data,
outcomes = "INO",
predictors = c("intervention", "intervention * time"),
covariates = c("", "+ gender + age"),
id = "pat_id"
)

}
