% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_log_mult.R
\name{glm_log_mult}
\alias{glm_log_mult}
\title{glm_log_mult: run multiple logistic regressions and output in table}
\usage{
glm_log_mult(
  data,
  outcomes,
  predictors,
  covariates = "",
  exponentiate = TRUE,
  show_progress = FALSE,
  std_odds = FALSE
)
}
\arguments{
\item{data}{Dataframe containing all variables for regression analyses.}

\item{outcomes}{A vector with your outcome variables (dependent variables), where the names of your variables are in quotes. For example: c(‘outcome1’, ‘outcome2’, ‘outcome3’).}

\item{predictors}{A vector containing your predictors (independent variables), where the names of your variables are in quotes. For example: c(‘predictor1’, ‘predictor2’, ‘predictor3’). It is possible to directly recode binary or categorical variables and get an output for both encodings, e.g. c(‘predictor1’, ‘fct_rev(predictor1)’). It is also possible to specify interaction terms, e.g. c(‘predictor1 * predictor2’, ‘predictor2 * predictor 3’).}

\item{covariates}{A vector containing your covariates (variables you want to correct for), where the entire enumeration of covariates per model are in quotes and different covariates in the same model are separated by a plus. Also, covariates should be preceded by a plus. For example: c(‘+ covariate 1 + covariate 2’, ‘+ covariate2 + covariate3’). If you want to run a model with and without covariates then specify it like this: c(‘’, ‘+ covariate 1 + covariate 2’)}

\item{exponentiate}{Do you want to exponentiate the output (e^x) so that you get the odds ratio? Default is TRUE.}

\item{show_progress}{Do you want to see the progress of running the models? Default is FALSE.}

\item{std_odds}{Do you want to standardize the log(odds)? Default is FALSE.}
}
\value{
Dataframe with results from multiple logistic regression analyses.
}
\description{
glm_log_mult: run multiple logistic regressions and output in table
}
\examples{

glm_log_mult(
data = MS_trial_data,
outcomes = "INO",
predictors = c("intervention", "age"),
covariates = c("", "+ gender")
)

}
