% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mult.R
\name{lm_mult}
\alias{lm_mult}
\title{lm_mult: Run multiple linear regression models and output table}
\usage{
lm_mult(
  data,
  outcomes,
  predictors,
  covariates = "",
  std_beta = FALSE,
  exp_log = FALSE,
  show_progress = FALSE
)
}
\arguments{
\item{data}{data frame with all your data.}

\item{outcomes}{a vector with your outcome variables (dependent variables), where the names of your variables are in quotes. For example: c(‘outcome1’, ‘outcome2’, ‘outcome3’).}

\item{predictors}{A vector containing your predictors (independent variables), where the names of your variables are in quotes. For example: c(‘predictor1’, ‘predictor2’, ‘predictor3’). It is possible to directly recode binary or categorical variables and get an output for both encodings, e.g. c(‘predictor1’, ‘fct_rev(predictor1)’). It is also possible to specify interaction terms, e.g. c(‘predictor1 * predictor2’, ‘predictor2 * predictor 3’).}

\item{covariates}{a vector containing your covariates (variables you want to correct for), where the entire enumeration of covariates per model are in quotes and different covariates in the same model are separated by a plus. Also, covariates should be preceded by a plus. For example: c(‘+ covariate 1 + covariate 2’, ‘+ covariate2 + covariate3’). If you want to run a model with and without covariates then specify it like this: c(‘’, ‘+ covariate 1 + covariate 2’)}

\item{std_beta}{Do you also want standardized betas in the output, set this to TRUE.}

\item{exp_log}{Do you want to exponentiate the results for log transformed outcomes, set this to TRUE.}

\item{show_progress}{Do you want to see the progress of the function, set this to TRUE.}
}
\value{
Large table with results from all regression formulas
}
\description{
These functions allow you to specify multiple outcome measures, predictors and covariates.
The function then creates regression formulas from all combinations of these variables.
The output is a large table with results from all regression formulas.
}
\examples{
lm_mult(
data = MS_trial_data,
outcomes = c("EDSS", "SDMT"),
predictors = "age",
covariates = c("", " + gender")
)

}
