% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_assump_tests.R
\name{lmer_assump_tests}
\alias{lmer_assump_tests}
\title{lmer_assump_tests: Generate plots for checking linear mixed model assumptions}
\usage{
lmer_assump_tests(model)
}
\arguments{
\item{model}{A linear mixed model object created with \code{lme4::lmer()}.}
}
\value{
Plots for checking linear mixed model assumptions: a histogram of residuals for normality, a residuals vs fitted plot for linearity, and a scale-location plot for homoscedasticity.
}
\description{
Generate plots for checking linear mixed model assumptions: normality of residuals, linearity, and homoscedasticity.
}
\examples{

model <- lme4::lmer(SDMT ~ age + (1|gender), data = MS_trial_data, REML = FALSE)
lmer_assump_tests(model)

}
