% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_mult_assump_tests.R
\name{lmer_mult_assump_tests}
\alias{lmer_mult_assump_tests}
\title{lmer_mult_assump_tests: Generate plots for checking linear mixed model assumptions for multiple models}
\usage{
lmer_mult_assump_tests(
  data,
  outcomes = NULL,
  predictors = NULL,
  covariates = "",
  randoms = NULL,
  formulas = NULL,
  control = lme4::lmerControl()
)
}
\arguments{
\item{data}{data frame with outcomes, predictors, covariates and variables for random effects in long format with a row for every observation of the outcome variables. See \link{MS_trial_data} for an example dataset.}

\item{outcomes}{a vector with outcome variables (dependent variables), where the names of the outcome variables are in quotes, e.g. \code{c("outcome1", "outcome2", "outcome3")}. It is possible to specify log transformation of the outcome variables, e.g. \code{c("log(outcome1)", "log(outcome2)")}.}

\item{predictors}{a vector with predictors (independent variables), where the names of the variables are in quotes, e.g. \code{c("predictor1", "predictor2", "predictor3")}. It is possible to directly recode binary or categorical variables and get an output for both encodings, e.g. \code{c("predictor1", "fct_rev(predictor1)")}. It is also possible to specify interaction terms, e.g. \code{c("predictor1 * predictor2", "predictor2 * predictor 3")}.}

\item{covariates}{a vector containing covariates (variables you want to correct for), where the entire enumeration of covariates are in quotes, different covariates in the same model are separated and preceded by a plus, e.g. \code{c("+ covariate 1 + covariate 2", "+ covariate2 + covariate3")}. If you want to run a model with and without covariates then specify it like this: \code{c("", "+ covariate 1 + covariate 2")}.}

\item{randoms}{a vector containing random effects, where the entire enumeration of random effects are in quotes, different random effects in the same model are separated and preceded by a plus and notation from the lme4 package used, e.g. \code{c("+ (1|var1)", "+ (var2|var1)", "+ (1|var1/var3)", "+ (var2|var1) + (1|var1:var3)", "+ (var2|var1/var3)")}. If you want to run a model with and without random effects, specify it like this: \code{c("", "+ (1|var1)")}.}

\item{formulas}{optional vector of formulas specified manually, e.g. \code{c("outcome1 ~ predictor1 + covariate1 + (1|var1)", "outcome2 ~ predictor2 + covariate2 + (1|var2)")}. You can combine this with using the other inputs for single variables, all models are then run.}

\item{control}{This gives you access to the ‘control’ input of the \code{\link[lme4:lmer]{lme4::lmer()}} function and \code{\link[lme4:lmerControl]{lme4::lmerControl()}}. This allows you to bypass certain errors. For example, in case of error about number of observations, you can try control = lmerControl(check.nobs.vs.nRE = ‘ignore’).}
}
\value{
A list of plots for checking linear mixed model assumptions for each model.
}
\description{
Runs multiple linear regression models, similar to \code{\link[=lmer_mult]{lmer_mult()}}, and passes each model to \code{\link[=lmer_assump_tests]{lmer_assump_tests()}} to generate plots for checking normality of residuals and linearity.
Plots for homoscedasticity currently not available.
}
\examples{

lmer_mult_assump_tests(
 data = MS_trial_data,
 outcomes = c("EDSS", "SDMT"),
 predictors = c("intervention", "intervention * time"),
 covariates = c("", "+ age + gender"),
 randoms = c("+ (1|pat_id)")
 )

}
